refactored general_components
